+++ src/schemes/contributionReward.ts
@@ -1,0 +1,24 @@
+import BN = require('bn.js')
+import { Address } from '../types'
+
+export interface IContributionReward {
+  beneficiary: Address
+  externalTokenReward: BN
+  externalToken: Address
+  ethReward: BN
+  nativeTokenReward: BN
+  periods: number
+  periodLength: number
+  reputationReward: BN
+}
+
+export interface IProposalCreateContributionRewardOptions {
+  beneficiary?: Address
+  nativeTokenReward?: BN
+  reputationReward?: BN
+  ethReward?: BN
+  externalTokenReward?: BN
+  externalTokenAddress?: Address
+  periodLength?: number
+  periods?: any
+}
