+++ src/schemes/schemeRegistrar.ts
@@ -1,0 +1,18 @@
+import { Address } from '../types'
+
+export interface ISchemeRegistrar {
+  id: string
+  schemeToRegister: Address
+  schemeToRegisterParamsHash: string
+  schemeToRegisterPermission: string
+  schemeToRemove: string
+  decision: number
+  schemeRegistered: boolean
+  schemeRemoved: boolean
+}
+
+export interface IProposalCreateOptions {
+  parametersHash?: string
+  permissions?: string
+  scheme?: Address
+}
